Rename setupbase for tab completion
It is now basesetup.py
